Logs_sepc fix
