+++ app/client/cypress/e2e/Regression_TestSuite/ClientSideTests/OtherUIFeatures/Logs2_spec.ts
@@ -1,4 +1,4 @@
-const commonlocators = require("../../../../locators/commonlocators.json");
+import commonlocators from "../../../../locators/commonlocators.json";
 import { ObjectsRegistry } from "../../../../support/Objects/Registry";
 
 const locator = ObjectsRegistry.CommonLocators;
@@ -12,8 +12,7 @@ const {
   PropertyPane: propPane,
 } = ObjectsRegistry;
 
-let logString: string;
-let dataSet: unknown;
+let logString, dataSet;
 
 const generateTestLogString = () => {
   const randString = Cypress._.random(0, 1e4);
@@ -31,149 +30,9 @@ describe("Debugger logs", function () {
     logString = generateTestLogString();
   });
 
-  it("1. Modifying widget properties should log the same", function () {
-    ee.DragDropWidgetNVerify("buttonwidget", 200, 200);
-    propPane.UpdatePropertyFieldValue("Label", "Test");
-    debuggerHelper.ClickDebuggerIcon(0, true, 0);
-    agHelper.GetNClick(jsEditor._logsTab);
-    debuggerHelper.LogStateContains("Test");
-  });
-
-  it("2. Reset debugger state", function () {
-    cy.get(".t--property-control-visible").find(".t--js-toggle").click();
-    cy.testJsontext("visible", "Test");
-    cy.get(commonlocators.homeIcon).click({ force: true });
-    cy.generateUUID().then((id) => {
-      cy.CreateAppInFirstListedWorkspace(id);
-      debuggerHelper.AssertErrorCount(0);
-    });
-  });
-
-  it("3. Console log on button click with normal moustache binding", function () {
-    ee.DragDropWidgetNVerify("buttonwidget", 200, 200);
-    // Testing with normal log in moustache binding
-    propPane.EnterJSContext("onClick", `{{console.log("${logString}")}}`);
-    agHelper.Sleep(2000);
-    agHelper.ClickButton("Submit");
-    debuggerHelper.ClickDebuggerIcon();
-    agHelper.GetNClick(jsEditor._logsTab);
-    debuggerHelper.DoesConsoleLogExist(logString);
-  });
-
-  it("4. Console log on button click with arrow function IIFE", function () {
-    debuggerHelper.ClearLogs();
-    ee.SelectEntityByName("Button1");
-    // Testing with normal log in iifee
-    propPane.EnterJSContext(
-      "onClick",
-      `{{(() => {
-          console.log('${logString}');
-        }) () }}`,
-    );
-    agHelper.ClickButton("Submit");
-    debuggerHelper.DoesConsoleLogExist(logString);
-  });
-
-  it("5. Console log on button click with function keyword IIFE", function () {
-    debuggerHelper.ClearLogs();
-    ee.SelectEntityByName("Button1");
-    // Testing with normal log in iifee
-    propPane.EnterJSContext(
-      "onClick",
-      `{{ function () {
-          console.log('${logString}');
-        } () }}`,
-    );
-    agHelper.ClickButton("Submit");
-    debuggerHelper.DoesConsoleLogExist(logString);
-  });
-
-  it("6. Console log on button click with async function IIFE", function () {
-    debuggerHelper.ClearLogs();
-    // Testing with normal log in iifee
-    ee.SelectEntityByName("Button1");
-    propPane.EnterJSContext(
-      "onClick",
-      `{{(async() => {
-          console.log('${logString}');
-        }) () }}`,
-    );
-    agHelper.ClickButton("Submit");
-    debuggerHelper.DoesConsoleLogExist(logString);
-  });
-
-  it("7. Console log on button click with mixed function IIFE", function () {
-    debuggerHelper.ClearLogs();
-    // Testing with normal log in iifee
-    ee.SelectEntityByName("Button1");
-    const logStringChild = generateTestLogString();
-    propPane.EnterJSContext(
-      "onClick",
-      `{{ function () {
-          console.log('${logString}');
-          (async () => {console.log('${logStringChild}')})();
-        } () }}`,
-    );
-    agHelper.ClickButton("Submit");
-    debuggerHelper.DoesConsoleLogExist(logString);
-    debuggerHelper.DoesConsoleLogExist(logStringChild);
-  });
-
-  it("8. Console log grouping on button click", function () {
-    debuggerHelper.ClearLogs();
-    // Testing with normal log in iifee
-    ee.SelectEntityByName("Button1");
-    propPane.EnterJSContext(
-      "onClick",
-      `{{ function () {
-          console.log('${logString}');
-          console.log('${logString}');
-          console.log('${logString}');
-          console.log('${logString}');
-          console.log('${logString}');
-        } () }}`,
-    );
-    agHelper.ClickButton("Submit");
-    debuggerHelper.DoesConsoleLogExist(logString);
-    debuggerHelper.Assert_Consecutive_Console_Log_Count(5);
-  });
-
-  it("9. Console log grouping on button click with different log in between", function () {
-    debuggerHelper.ClearLogs();
-    // Testing with normal log in iifee
-    ee.SelectEntityByName("Button1");
-    propPane.EnterJSContext(
-      "onClick",
-      `{{ function () {
-          console.log('${logString}');
-          console.log('${logString}');
-          console.log('Different ${logString}');
-          console.log('${logString}');
-        } () }}`,
-    );
-    agHelper.ClickButton("Submit");
-    debuggerHelper.DoesConsoleLogExist(logString);
-    debuggerHelper.DoesConsoleLogExist(`Different ${logString}`);
-    debuggerHelper.Assert_Consecutive_Console_Log_Count(2);
-  });
-
-  it("10. Console log grouping on button click from different source", function () {
-    debuggerHelper.ClearLogs();
-    // Testing with normal log in iifee
-    ee.SelectEntityByName("Button1");
-    propPane.EnterJSContext("onClick", `{{console.log("${logString}")}}`);
-    // Add another button
-    ee.DragDropWidgetNVerify("buttonwidget", 400, 200);
-    propPane.UpdatePropertyFieldValue("Label", "Submit2");
-    propPane.EnterJSContext("onClick", `{{console.log("${logString}")}}`);
-    agHelper.Sleep(2000);
-    agHelper.ClickButton("Submit");
-    agHelper.ClickButton("Submit2");
-    debuggerHelper.DoesConsoleLogExist(logString);
-    debuggerHelper.Assert_Consecutive_Console_Log_Count(0);
-  });
-
-  it("11. Console log on text widget with normal moustache binding", function () {
+  it("1. Console log on text widget with normal moustache binding", function () {
+    ee.NavigateToSwitcher("Widgets");
+    agHelper.TypeText(locator._entityExplorersearch, "Text");
     ee.DragDropWidgetNVerify("textwidget", 400, 400);
     propPane.UpdatePropertyFieldValue(
       "Text",
@@ -192,7 +51,7 @@ describe("Debugger logs", function () {
     debuggerHelper.DoesConsoleLogExist(logString);
   });
 
-  it("12. Console log in sync function", function () {
+  it("2. Console log in sync function", function () {
     ee.NavigateToSwitcher("Explorer");
     jsEditor.CreateJSObject(
       `export default {
@@ -216,7 +75,7 @@ describe("Debugger logs", function () {
     debuggerHelper.DoesConsoleLogExist(logString);
   });
 
-  it("13. Console log in async function", function () {
+  it("3. Console log in async function", function () {
     ee.NavigateToSwitcher("Explorer");
     jsEditor.CreateJSObject(
       `export default {
@@ -255,7 +114,7 @@ describe("Debugger logs", function () {
     debuggerHelper.DoesConsoleLogExist(logString);
   });
 
-  it("14. Console log after API succedes", function () {
+  it("4. Console log after API succedes", function () {
     ee.NavigateToSwitcher("Explorer");
     apiPage.CreateAndFillApi(dataSet.baseUrl + dataSet.methods, "Api1");
     const returnText = "success";
@@ -287,8 +146,8 @@ describe("Debugger logs", function () {
 
     cy.get("@jsObjName").then((jsObjName) => {
       agHelper.Sleep(2000);
-      agHelper.GetNClick(jsEditor._runButton);
-      agHelper.GetNClick(jsEditor._logsTab, 0, true);
+      jsEditor.RunJSObj();
+      agHelper.GetNClick(jsEditor._logsTab);
       debuggerHelper.DoesConsoleLogExist(`${logString} Started`);
       debuggerHelper.DoesConsoleLogExist(`${logString} Success`);
       ee.DragDropWidgetNVerify("textwidget", 200, 600);
@@ -302,7 +161,7 @@ describe("Debugger logs", function () {
     });
   });
 
-  it("15. Console log after API execution fails", function () {
+  it("5. Console log after API execution fails", function () {
     ee.NavigateToSwitcher("Explorer");
     apiPage.CreateAndFillApi(dataSet.baseUrl + dataSet.methods + "xyz", "Api2");
     jsEditor.CreateJSObject(
@@ -328,13 +187,13 @@ describe("Debugger logs", function () {
         shouldCreateNewJSObj: true,
       },
     );
-    agHelper.GetNClick(jsEditor._runButton);
+    jsEditor.RunJSObj();
     agHelper.GetNClick(jsEditor._logsTab);
     debuggerHelper.DoesConsoleLogExist(`${logString} Started`);
     debuggerHelper.DoesConsoleLogExist(`${logString} Failed`);
   });
 
-  it("16. Console log source inside nested function", function () {
+  it("6. Console log source inside nested function", function () {
     jsEditor.CreateJSObject(
       `export default {
         myFun1: async () => {
@@ -352,13 +211,13 @@ describe("Debugger logs", function () {
         shouldCreateNewJSObj: false,
       },
     );
-    agHelper.GetNClick(jsEditor._runButton);
+    jsEditor.RunJSObj();
     agHelper.GetNClick(jsEditor._logsTab);
     debuggerHelper.DoesConsoleLogExist(`Parent ${logString}`);
     debuggerHelper.DoesConsoleLogExist(`Child ${logString}`);
   });
 
-  it("17. Console log grouping", function () {
+  it("7. Console log grouping", function () {
     jsEditor.CreateJSObject(
       `export default {
         myFun1: async () => {
@@ -379,13 +238,13 @@ describe("Debugger logs", function () {
         shouldCreateNewJSObj: true,
       },
     );
-    agHelper.GetNClick(jsEditor._runButton);
+    jsEditor.RunJSObj();
     agHelper.GetNClick(jsEditor._logsTab);
     debuggerHelper.DoesConsoleLogExist(`${logString}`);
     debuggerHelper.Assert_Consecutive_Console_Log_Count(5);
   });
 
-  it("18. Console log should not mutate the passed object", function () {
+  it("8. Console log should not mutate the passed object", function () {
     ee.NavigateToSwitcher("Explorer");
     jsEditor.CreateJSObject(
       `export default {
@@ -414,7 +273,7 @@ describe("Debugger logs", function () {
     debuggerHelper.DoesConsoleLogExist("end: [0,1,2,3,4]");
   });
 
-  it("19. Bug #19115 - Objects that start with an underscore `_JSObject1` fail to be navigated from the debugger", function () {
+  it("9. Bug #19115 - Objects that start with an underscore `_JSObject1` fail to be navigated from the debugger", function () {
     const JSOBJECT_WITH_UNNECCESARY_SEMICOLON = `export default {
         myFun1: () => {
             //write code here
